+++ src/main/scala/sc4pac/Find.scala
@@ -81,7 +81,7 @@ object Find {
 
     def pickVariant(pkgOpt: Option[JD.Package]): Task[(JD.Package, JD.VariantData)] = pkgOpt match {
       case None => ZIO.fail(new Sc4pacVersionNotFound(s"Could not find metadata of ${module.orgName} or suitable variant.",
-                                                      "Either the package name is spelled incorrectly or the metadata stored in the corresponding channel is incorrect or incomplete."))
+                                                      "Either the package name is spelled incorrectly or the metadata stored in the corresponding channel is incorrect or incomplete.", module))
       case Some(pkgData) =>
         pkgData.variants.find(vd => isSubMap(vd.variant, globalVariant)) match {
           case Some(vd) => ZIO.succeed((pkgData, vd))