+++ src/main/scala/sc4pac/api/api.scala
@@ -41,6 +41,7 @@ class Api(options: sc4pac.cli.Commands.ServerOptions) {
 
   private def expectedFailureMessage(err: cli.Commands.ExpectedFailure): ErrorMessage = err match {
     case abort: error.Sc4pacVersionNotFound => ErrorMessage.VersionNotFound(abort.title, abort.detail)
+    case abort: error.UnresolvableDependencies => ErrorMessage.UnresolvableDependencies(abort.title, abort.detail)
     case abort: error.Sc4pacAssetNotFound => ErrorMessage.AssetNotFound(abort.title, abort.detail)
     case abort: error.ExtractionFailed => ErrorMessage.ExtractionFailed(abort.title, abort.detail)
     case abort: error.UnsatisfiableVariantConstraints => ErrorMessage.UnsatisfiableVariantConstraints(abort.title, abort.detail)
@@ -56,6 +57,7 @@ class Api(options: sc4pac.cli.Commands.ServerOptions) {
 
   private def expectedFailureStatus(err: cli.Commands.ExpectedFailure): Status = err match {
     case abort: error.Sc4pacVersionNotFound => Status.NotFound
+    case abort: error.UnresolvableDependencies => Status.NotFound
     case abort: error.Sc4pacAssetNotFound => Status.NotFound
     case abort: error.ExtractionFailed => Status.InternalServerError
     case abort: error.UnsatisfiableVariantConstraints => Status.InternalServerError