+++ src/main/scala/sc4pac/api/message.scala
@@ -40,6 +40,20 @@ object PromptMessage {
   val yesNo = Seq("Yes", "No")
   val yes = yesNo.head
 
+  @upickle.implicits.key("/prompt/confirmation/update/remove-unresolvable-packages")
+  case class ConfirmRemoveUnresolvablePackages(
+    packages: Seq[BareModule],
+    choices: Seq[String], // = yesNo,
+    token: String,
+    responses: Map[String, ResponseMessage]
+  ) extends PromptMessage derives UP.ReadWriter
+  object ConfirmRemoveUnresolvablePackages {
+    def apply(packages: Seq[BareModule]): ConfirmRemoveUnresolvablePackages = {
+      val token = scala.util.Random.nextInt().toHexString
+      ConfirmRemoveUnresolvablePackages(packages = packages, choices = yesNo, token = token, responsesFromChoices(yesNo, token))
+    }
+  }
+
   @upickle.implicits.key("/prompt/confirmation/update/plan")
   case class ConfirmUpdatePlan(
     toRemove: Seq[ConfirmUpdatePlan.Pkg],
@@ -96,6 +110,8 @@ object ErrorMessage {
   case class ReadingProfileFailed(title: String, detail: String) extends ErrorMessage derives UP.ReadWriter
   @upickle.implicits.key("/error/version-not-found")
   case class VersionNotFound(title: String, detail: String) extends ErrorMessage derives UP.ReadWriter
+  @upickle.implicits.key("/error/unresolvable-dependencies")
+  case class UnresolvableDependencies(title: String, detail: String) extends ErrorMessage derives UP.ReadWriter
   @upickle.implicits.key("/error/asset-not-found")
   case class AssetNotFound(title: String, detail: String) extends ErrorMessage derives UP.ReadWriter
   @upickle.implicits.key("/error/extraction-failed")