+++ src/main/scala/sc4pac/error.scala
@@ -16,7 +16,9 @@ final class Sc4pacPublishIncomplete(val title: String, val detail: String) exten
 
 final class Sc4pacAssetNotFound(val title: String, val detail: String) extends java.io.IOException(s"$title $detail") with Sc4pacErr
 
-final class Sc4pacVersionNotFound(val title: String, val detail: String) extends java.io.IOException(s"$title $detail") with Sc4pacErr
+final class Sc4pacVersionNotFound(val title: String, val detail: String, val dep: BareDep) extends java.io.IOException(s"$title $detail") with Sc4pacErr
+
+final class UnresolvableDependencies(val title: String, val detail: String, val deps: Seq[BareDep]) extends java.io.IOException(s"$title $detail") with Sc4pacErr
 
 final class Sc4pacMissingVariant(val packageData: JsonData.Package, msg: String) extends scala.util.control.ControlThrowable(msg)